+++ src/react/wrappers/withErrorBoundary.tsx
@@ -1,5 +1,6 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { DevErrorBoundary } from "../DevErrorBoundary";
+import { createPortal } from "react-dom";
 
 export function withErrorBoundary(
   appDirectory: string,
@@ -8,8 +9,27 @@ export function withErrorBoundary(
   return () => {
     const [intercept, setIntercept] = useState(true);
 
+    const [container, setContainer] = useState<HTMLDivElement | null>(null);
+
+    useEffect(() => {
+      const container = document.createElement("div");
+      // Append to the end of the body
+      document.body.appendChild(container);
+      setContainer(container);
+    }, []);
+
     if (intercept)
-      return (
+      return container ? (
+        createPortal(
+          <DevErrorBoundary
+            appDirectory={appDirectory}
+            onRenderOriginalErrorBoundary={
+              OriginalErrorBoundary ? () => setIntercept(false) : undefined
+            }
+          />,
+          container
+        )
+      ) : (
         <DevErrorBoundary
           appDirectory={appDirectory}
           onRenderOriginalErrorBoundary={